src/main/java/inu/amigo/orderIt/service/MemberService.java
package inu.amigo.orderIt.service;


import inu.amigo.orderIt.domain.user.Member;
import inu.amigo.orderIt.repository.MemberRepository;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.lang.Nullable;
import org.springframework.stereotype.Service;

import java.util.List;

@Service
public class MemberService {

    @Autowired
    MemberRepository memberRepository;

    /**
     * 회원 가입
     */
    public Long join(Member member) {
        validateDuplicateMember(member);
        memberRepository.save(member);
        return member.getMemberId();
    }

    private void validateDuplicateMember(Member member) {
        List<Member> findMembers = memberRepository.findByUsername(member.getUsername());
        if (!findMembers.isEmpty()) {
            throw new IllegalStateException("이미 존재하는 회원입니다.");
        }
    }

    /**
     * 전체 회원 조회
     */
    public Member findOne(Long memberId) {
        return memberRepository.findById(memberId).orElse(new Member());
    }
}
